Fix: MapZone Loading, Valid Instancing & Adapter
Fixed Map Zones loading on logging in.

Added a method to check if the Players current MapZone save is valid instance of the MapZone class. This removes the "instanceof" calling and the exclusive casting for the target class to inherit their methods. (Cleaner and easier executions).

Added a json adapter class for MapZone.